fix(leaderboard): harden query ints and row type placement
