+++ api/services/database/database.ts
@@ -16,20 +16,11 @@ import type { LeaderboardSeriesRow } from "./types/leaderboard_series";
 import type { LeaderboardSeriesPlayersRow } from "./types/leaderboard_series_players";
 import type { LeaderboardGamesRow } from "./types/leaderboard_games";
 import type { LeaderboardGamePlayersRow } from "./types/leaderboard_game_players";
+import type { LeaderboardRankingRow } from "./types/leaderboard_ranking_row";
 import type { LeaderboardConfigRow } from "./types/leaderboard_config";
 import { LeaderboardWindow, LeaderboardMetric } from "./types/leaderboard_config";
 
 const DEFAULT_LEADERBOARD_ENABLED_WINDOWS_JSON = '["1W","1M","3M","6M","12M"]';
-
-export interface LeaderboardRankingRow {
-  XboxXuid: string;
-  DiscordUserId: string | null;
-  Gamertag: string;
-  SeriesPlayed: number;
-  SeriesWins: number;
-  GamesPlayed: number;
-  MetricValue: number;
-}
 
 interface LeaderboardRankingsQuery {
   guildId: string;
