+++ api/services/database/types/leaderboard_ranking_row.ts
@@ -1,0 +1,9 @@
+export interface LeaderboardRankingRow {
+  XboxXuid: string;
+  DiscordUserId: string | null;
+  Gamertag: string;
+  SeriesPlayed: number;
+  SeriesWins: number;
+  GamesPlayed: number;
+  MetricValue: number;
+}
